# specs/001-association-hub/contracts/api.yaml
openapi: 3.0.0
info:
  title: AssociationHub MVP API
  description: REST API for AssociationHub member management and email communication platform
  version: 1.0.0
  contact:
    name: AssociationHub
    url: https://github.com/associationhub/associationhub
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://your-deployment.vercel.app/api
    description: Production
...
paths:
...
  /templates/select:
    post:
      tags:
        - Templates
      summary: Select association template
      description: Select template on first login (parents, sports, cultural)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - template
              properties:
                template:
                  type: string
                  enum: [parents, sports, cultural]
                  example: parents
      responses:
        '200':
          description: Template selected and database seeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  template:
                    type: string
                    example: parents
                  roles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/GroupType'
        '400':
          description: Invalid template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /members/import:
    post:
      tags:
        - Members
      summary: Import members from CSV
      description: Bulk import members from CSV file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with columns name,email,role,group
      responses:
        '200':
          description: Import completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: integer
                    example: 45
                  failed:
                    type: integer
                    example: 5
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        row:
                          type: integer
                        error:
                          type: string
        '400':
          description: Invalid CSV format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /emails/send:
    post:
      tags:
        - Emails
      summary: Send email campaign
      description: Send email to filtered members with optional attachment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subject
                - body
              properties:
                subject:
                  type: string
                  example: Réunion importante
                body:
                  type: string
                  example: Bonjour {Nom}, vous êtes invité à la réunion.
                attachmentUrl:
                  type: string
                  format: uri
                  example: https://blob.vercel-storage.com/file.pdf
                roleId:
                  type: string
                  description: Filter recipients by role
                groupId:
                  type: string
                  description: Filter recipients by group
      responses:
        '200':
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  recipientCount:
                    type: integer
                    example: 45
                  campaignId:
                    type: string
                    example: clxyz123
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: string
          example: clxyz123
        name:
          type: string
          example: delegue_titulaire
        displayName:
          type: string
          example: Délégué titulaire
        sortOrder:
          type: integer
          example: 1
    GroupType:
      type: object
      properties:
        id:
          type: string
          example: clxyz123
        name:
          type: string
          example: 6e1
        category:
          type: string
          example: classe
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid credentials
        details:
          type: object
          nullable: true
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
security:
  - sessionAuth: []